Agent/requirements.txt
python-dotenv
smolagents
openai
